# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/beacon": {
            "get": {
                "description": "describes the current status of beacond",
                "produces": [
                    "application/json"
                ],
                "summary": "Get beacon details",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.BeaconDescribeResponse"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "reports the health of the beacond server",
                "produces": [
                    "application/json"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.BaseResponse"
                        }
                    }
                }
            }
        },
        "/probe": {
            "post": {
                "description": "creates a probe for the namespace and repo provided in the URL query parameters",
                "produces": [
                    "application/json"
                ],
                "summary": "Create a probe",
                "parameters": [
                    {
                        "type": "string",
                        "description": "the repo namespace the probe should check for image updates",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "the repo name which the probe should check for image updates",
                        "name": "repo",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/server.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.BaseResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/server.BaseResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/server.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.BaseResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "deletes the probe for the namespace and repo provided in the URL query parameters",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a probe",
                "parameters": [
                    {
                        "type": "string",
                        "description": "the repo namespace the probe should check for image updates",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "the repo name which the probe should check for image updates",
                        "name": "repo",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/server.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.BaseResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/server.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.BaseResponse"
                        }
                    }
                }
            }
        },
        "/probes": {
            "get": {
                "description": "lists probes that are running for beacond",
                "produces": [
                    "application/json"
                ],
                "summary": "Lists all probes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.ListProbesResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "server.BaseResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "server.BeaconDescribeResponse": {
            "type": "object",
            "properties": {
                "probes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "registry": {
                    "type": "string"
                },
                "runtime": {
                    "type": "string"
                }
            }
        },
        "server.ListProbesResponse": {
            "type": "object",
            "properties": {
                "probes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    }
}